requirements.txt
Flask
imutils
opencv.python.headless==4.2.0.32
Pillow
scikit-image
gunicorn
